wip: Moved main to lib
